requirements.txt
Flask==2.3.3
spotipy==2.23.0
pandas==2.1.1
python-dotenv==1.0.0
requests==2.31.0 